+++ src/main/kotlin/de/mariushoefler/flutterenhancementsuite/exceptions/PubExceptions.kt
@@ -8,6 +8,6 @@ class GetLatestPackageVersionException(p: String) :
 class GetCurrentPackageVersionException(p: String) :
     Exception("Cannot read current version number for package: $p")
 
-class PubApiCouldNotBeReached(p: Exception) : Exception("Pub api could not be reached: $p")
+class PubApiCouldNotBeReachedException(p: Exception) : Exception("Pub api could not be reached: $p")
 
-class PubApiUnknownFormat(p: JsonSyntaxException) : Exception("Unexpected response from Pub api: $p")
+class PubApiUnknownFormatException(p: JsonSyntaxException) : Exception("Unexpected response from Pub api: $p")